refactor: gqls, responseModels utils로 옮김
