+++ frontend/src/utils/gqls.ts
@@ -1,5 +1,118 @@
 import gql from 'graphql-tag';
 
+export const GET_ALL_CHANNELS = gql`
+  query GetAllChannels($limit: Float!, $offset: Float!) {
+    channels(limit: $limit, offset: $offset) {
+      id
+      title
+      is_private
+      owner {
+        nickname
+      }
+      participants {
+        nickname
+      }
+    }
+  }
+`;
+
+export const ADD_CHANNEL = gql`
+  mutation AddChannel(
+    $owner_user_id: String!
+    $title: String!
+    $password: String
+  ) {
+    addChannel(
+      owner_user_id: $owner_user_id
+      title: $title
+      password: $password
+    ) {
+      id
+      title
+      is_private
+      owner {
+        id
+        nickname
+        # avatar
+        # status
+      }
+      administrators {
+        id
+        nickname
+        # avatar
+        # status
+      }
+      participants {
+        id
+        nickname
+        # avatar
+        # status
+      }
+    }
+  }
+`;
+// TODO: 현재 아바타, 스테이터스 불러오면 에러 발생하는 백엔드 오류 있어 주석 처리
+
+export const GET_CURRENT_CHANNEL = gql`
+  query GetCurrentChannel($id: ID!) {
+    user(id: $id) {
+      channel {
+        id
+        title
+        is_private
+        owner {
+          id
+          nickname
+          # avatar
+          # status
+        }
+        administrators {
+          id
+          nickname
+          # avatar
+          # status
+        }
+        participants {
+          id
+          nickname
+          # avatar
+          # status
+        }
+        bannedUsers {
+          id
+          nickname
+        }
+        mutedUsers {
+          id
+          nickname
+        }
+      }
+    }
+  }
+`;
+
+export const ENTER_CHANNEL = gql`
+  mutation EnterChannel($channel_id: ID!, $user_id: ID!) {
+    enterChannel(channel_id: $channel_id, user_id: $user_id) {
+      id
+      title
+      is_private
+      owner {
+        id
+        nickname
+      }
+      administrators {
+        id
+        nickname
+      }
+      participants {
+        id
+        nickname
+      }
+    }
+  }
+`;
+
 export const CHAT_MESSAGE = gql`
   mutation ChatMessage($message: String!, $user_id: ID!, $channel_id: ID!) {
     chatMessage(message: $message, user_id: $user_id, channel_id: $channel_id)
