+++ frontend/src/utils/responseModels.ts
@@ -1,4 +1,4 @@
-import { IChannelNotify, IUser } from '../../utils/models';
+import { IChannel, IChannelListItem, IChannelNotify, IUser } from './models';
 
 export interface SubscribeChannelResponse {
   subscribeChannel: IChannelNotify;
@@ -23,3 +23,29 @@ export interface GetChannelMutedUsers {
 export interface GetChannelBannedUsers {
   user: { channel: { bannedUsers: { id: string; nickname: string } } };
 } // TODO: 스네이크 케이스로 변경 예정
+
+export interface AddChannelResponse {
+  addChannel: IChannel;
+}
+
+export interface GetCurrentChannelResponse {
+  user: {
+    channel: IChannel;
+  };
+}
+
+export interface GetAllChannelsResponse {
+  channels: IChannelListItem[];
+}
+
+export interface GetCurrentChannelIdResponse {
+  user: {
+    channel: { id: string };
+  };
+}
+
+export interface EnterChannelResponse {
+  enterChannel: {
+    channel: IChannel;
+  };
+}
